Refactor with BetaSwitchUrlCard to fix GitHub unused var build issue.
