+++ js/src/setup-mc/setup-stepper/setup-accounts/google-mc-account/switch-url-card/index.js
@@ -16,7 +16,6 @@ import useApiFetchCallback from '.~/hooks/useApiFetchCallback';
 import { useAppDispatch } from '.~/data';
 import ContentButtonLayout from '.~/components/content-button-layout';
 import './index.scss';
-import betaExistingProductListingsStatement from '../betaExistingProductListingsStatement';
 
 /**
  * Temporarily unused for beta testing period. This should be used in production later.
@@ -99,47 +98,4 @@ const SwitchUrlCard = ( props ) => {
 	);
 };
 
-/**
- * This is used temporarily for beta testing purpose. For production roll out, we should remove this and use the above SwitchUrlCard instead.
- *
- * @param {Object} props Props.
- */
-const BetaSwitchUrlCard = ( props ) => {
-	const { id, message, onSelectAnotherAccount = () => {} } = props;
-
-	const handleUseDifferentMCClick = () => {
-		onSelectAnotherAccount();
-	};
-
-	return (
-		<Section.Card className="gla-switch-url-card">
-			<Section.Card.Body>
-				<ContentButtonLayout>
-					<Subsection.Title>{ toAccountText( id ) }</Subsection.Title>
-				</ContentButtonLayout>
-				<ContentButtonLayout>
-					<div>
-						<Subsection.Title>{ message }</Subsection.Title>
-						<Subsection.HelperText>
-							{ betaExistingProductListingsStatement }
-						</Subsection.HelperText>
-					</div>
-				</ContentButtonLayout>
-			</Section.Card.Body>
-			<Section.Card.Footer>
-				<AppTextButton
-					isSecondary
-					onClick={ handleUseDifferentMCClick }
-				>
-					{ __(
-						'Or, use a different Merchant Center account',
-						'google-listings-and-ads'
-					) }
-				</AppTextButton>
-			</Section.Card.Footer>
-		</Section.Card>
-	);
-};
-
-// export default SwitchUrlCard;
-export default BetaSwitchUrlCard;
+export default SwitchUrlCard;
